processing-services/ocr-service/requirements.txt
pytesseract==0.3.10
opencv-python-headless==4.10.0.84
httpx==0.27.0
orjson==3.10.7
Pillow==10.4.0